move logic from interpreter.py to dwelling.py
